Changed print statements and pool size

+++ axfr-test.py
@@ -16,14 +16,18 @@ def checkaxfr(domain):
       nameserver = str(ns)[:-1]
       if nameserver is None or nameserver == "":
         continue
+
+      if os.path.exists("." + os.sep + DATAPATH + os.sep + domain + "-" + nameserver + ".zone"):
+        continue
+
       try:
         axfr = dns.query.xfr(nameserver, domain, lifetime=5)
-        print(domain + "@" + nameserver)
         try:
           zone = dns.zone.from_xfr(axfr)
           if zone is None:
             continue
           fHandle = open("." + os.sep + DATAPATH + os.sep + domain + "-" + nameserver + ".zone", "w")
+          print("Success: " + domain + " @ " + nameserver)
           for name, node in zone.nodes.items():
             rdatasets = node.rdatasets
             for rdataset in rdatasets:
@@ -34,10 +38,11 @@ def checkaxfr(domain):
       except Exception as e:
         continue
   except Exception as e:
-    return None
+    pass
+  print("Finished: " + domain)
 
 def main():
-  pool = Pool(processes=100)
+  pool = Pool(processes=20)
   lines = open("domains.txt", "r").readlines()
   pool.map(checkaxfr, lines)
 if __name__ == '__main__':
